requirements.txt
streamlitpytesseractopencv-python-headlessPillownumpynltkscikit-learnPyMuPDF==1.22.0langdetecttransformers==4.30.2PyPDF2protobuf==3.20.3